avoid unsigned/signed comparison warning

+++ src/Firmware/board/encoders/v2.cpp
@@ -11,17 +11,17 @@ int8_t Board::getEncoderState(uint8_t encoderNumber)
     }
 
     if (encInverted[encoderNumber])
-        return pulses*-1;
+    return pulses*-1;
 
     return pulses;
 }
 
 bool Board::encoderEnabled(uint8_t encoderNumber)
 {
-    for (int i=0; i<sizeof(encoderMap); i++)
+    for (int i=0; i<(int)sizeof(encoderMap); i++)
     {
         if (encoderMap[i] == encoderNumber)
-            return true;
+        return true;
     }
 
     return false;
